Move join transform boundedness check to FlinkTableBuilder
